feat: adapt send type for feishu alert
