+++ dolphinscheduler-alert/dolphinscheduler-alert-plugins/dolphinscheduler-alert-feishu/src/main/java/org/apache/dolphinscheduler/plugin/alert/feishu/FeiShuParamsConstants.java
@@ -20,25 +20,32 @@ package org.apache.dolphinscheduler.plugin.alert.feishu;
 public final class FeiShuParamsConstants {
 
     public static final String FEI_SHU_PROXY_ENABLE = "$t('isEnableProxy')";
-    static final String WEB_HOOK = "$t('webhook')";
-    static final String NAME_WEB_HOOK = "WebHook";
+    static final String WEB_HOOK_OR_APP_ID = "$t('webhookOrAppId')";
+    static final String NAME_WEB_HOOK_OR_APP_ID = "WebhookOrAppId";
     static final String NAME_FEI_SHU_PROXY_ENABLE = "IsEnableProxy";
-
     static final String FEI_SHU_PROXY = "$t('proxy')";
-
     static final String NAME_FEI_SHU_PROXY = "Proxy";
-
     static final String FEI_SHU_PORT = "$t('port')";
-
     static final String NAME_FEI_SHU_PORT = "Port";
-
     static final String FEI_SHU_USER = "$t('user')";
-
     static final String NAME_FEI_SHU_USER = "User";
-
     static final String FEI_SHU_PASSWORD = "$t('password')";
-
     static final String NAME_FEI_SHU_PASSWORD = "Password";
+    static final String PERSONAL_WORK_APP_SECRET = "$t('appSecret')";
+    static final String NAME_PERSONAL_WORK_APP_SECRET = "appSecret";
+    static final String OPEN_ID = "open_id";
+    static final String USER_ID = "user_id";
+    static final String UNION_ID = "union_id";
+    static final String EMAIL = "email";
+    static final String CHAT_ID = "chat_id";
+    static final String RECEIVE_ID = "$t('receiveId')";
+    static final String NAME_RECEIVE_ID = "receiveId";
+    static final String RECEIVE_ID_TYPE = "$t('receiveIdType')";
+    static final String NAME_RECEIVE_ID_TYPE = "receiveIdType";
+    static final String NAME_FEI_SHU_SEND_TYPE = "sendType";
+    static final String FEI_SHU_SEND_TYPE = "$t('sendType')";
+    static final String accessTokenUrl = "https://open.feishu.cn/open-apis/auth/v3/tenant_access_token/internal";
+    static final String accessMessageUrl = "https://open.feishu.cn/open-apis/im/v1/messages";
 
     private FeiShuParamsConstants() {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
