+++ dolphinscheduler-alert/dolphinscheduler-alert-plugins/dolphinscheduler-alert-feishu/src/main/java/org/apache/dolphinscheduler/plugin/alert/feishu/FeiShuPersonalWorkResponse.java
@@ -1,0 +1,33 @@
+package org.apache.dolphinscheduler.plugin.alert.feishu;
+
+final class FeiShuPersonalWorkResponse {
+
+    private int code;
+    private String msg;
+    private Object data;
+
+    public void setCode(int code) {
+        this.code = code;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public void setMsg(String msg) {
+        this.msg = msg;
+    }
+
+    public String getMsg() {
+        return msg;
+    }
+
+    public void setData(Object data) {
+        this.data = data;
+    }
+
+    public Object getData() {
+        return data;
+    }
+
+}
