+++ dolphinscheduler-alert/dolphinscheduler-alert-plugins/dolphinscheduler-alert-feishu/src/test/java/org/apache/dolphinscheduler/plugin/alert/feishu/FeiShuAlertChannelFactoryTest.java
@@ -33,7 +33,7 @@ public class FeiShuAlertChannelFactoryTest {
         FeiShuAlertChannelFactory feiShuAlertChannelFactory = new FeiShuAlertChannelFactory();
         List<PluginParams> params = feiShuAlertChannelFactory.params();
         JSONUtils.toJsonString(params);
-        Assertions.assertEquals(6, params.size());
+        Assertions.assertEquals(10, params.size());
     }
 
     @Test
